# Faster `changebasis` for `UniformBSplineSpace`