@hyrodium 2022-07-20
**current master**
```julia
julia> using BenchmarkTools, BasicBSpline

julia> k1 = UniformKnotVector(1:2:16)
UniformKnotVector(1:2:15)

julia> k2 = UniformKnotVector(1:16)
UniformKnotVector(1:16)

julia> P1 = UniformBSplineSpace{2}(k1)
UniformBSplineSpace{2, Int64, StepRange{Int64, Int64}}(UniformKnotVector(1:2:15))

julia> P2 = UniformBSplineSpace{2}(k2)
UniformBSplineSpace{2, Int64, UnitRange{Int64}}(UniformKnotVector(1:16))

julia> P1 ⊆ P2
true

julia> changebasis(P1,P2)
5×13 Matrix{Float64}:
 0.25  0.75  0.75  0.25  0.0   0.0   0.0   0.0   0.0   0.0   0.0   0.0   -0.0
 0.0   0.0   0.25  0.75  0.75  0.25  0.0   0.0   0.0   0.0   0.0   0.0   -0.0
 0.0   0.0   0.0   0.0   0.25  0.75  0.75  0.25  0.0   0.0   0.0   0.0   -0.0
 0.0   0.0   0.0   0.0   0.0   0.0   0.25  0.75  0.75  0.25  0.0   0.0   -0.0
 0.0   0.0   0.0   0.0   0.0   0.0   0.0   0.0   0.25  0.75  0.75  0.25  -0.0

julia> @benchmark changebasis(P1,P2)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  43.171 μs …  2.295 ms  ┊ GC (min … max): 0.00% … 96.24%
 Time  (median):     45.887 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   49.981 μs ± 56.375 μs  ┊ GC (mean ± σ):  2.89% ±  2.53%

  ▂▆██▇▆▄▃▃▃▃▂▃▃▃▄▃▃▁▁ ▁▁                                     ▂
  █████████████████████████▇█▇▆▆▅▅▄▄▆▅▅▆▁▅▆▅▇▇▅▆▇▆▇█▇▇▇▇▇▇▇▇█ █
  43.2 μs      Histogram: log(frequency) by time      84.5 μs <

 Memory estimate: 35.08 KiB, allocs estimate: 339.
```

**With this PR**
```julia
julia> using BenchmarkTools, BasicBSpline
[ Info: Precompiling BasicBSpline [4c5d9882-2acf-4ea4-9e48-968fd4518195]

julia> k1 = UniformKnotVector(1:2:16)
UniformKnotVector(1:2:15)

julia> k2 = UniformKnotVector(1:16)
UniformKnotVector(1:16)

julia> P1 = UniformBSplineSpace{2}(k1)
UniformBSplineSpace{2, Int64, StepRange{Int64, Int64}}(UniformKnotVector(1:2:15))

julia> P2 = UniformBSplineSpace{2}(k2)
UniformBSplineSpace{2, Int64, UnitRange{Int64}}(UniformKnotVector(1:16))

julia> P1 ⊆ P2
true

julia> changebasis(P1,P2)
5×13 Matrix{Float64}:
 0.25  0.75  0.75  0.25  0.0   0.0   0.0   0.0   0.0   0.0   0.0   0.0   0.0
 0.0   0.0   0.25  0.75  0.75  0.25  0.0   0.0   0.0   0.0   0.0   0.0   0.0
 0.0   0.0   0.0   0.0   0.25  0.75  0.75  0.25  0.0   0.0   0.0   0.0   0.0
 0.0   0.0   0.0   0.0   0.0   0.0   0.25  0.75  0.75  0.25  0.0   0.0   0.0
 0.0   0.0   0.0   0.0   0.0   0.0   0.0   0.0   0.25  0.75  0.75  0.25  0.0

julia> @benchmark changebasis(P1,P2)
BenchmarkTools.Trial: 10000 samples with 194 evaluations.
 Range (min … max):  511.845 ns …   9.821 μs  ┊ GC (min … max): 0.00% … 93.70%
 Time  (median):     532.660 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   597.090 ns ± 571.843 ns  ┊ GC (mean ± σ):  7.29% ±  7.09%

  ▂▇██▆▅▃▃▃▂▁▁▁▁   ▁▂▃▂▂                                        ▂
  ████████████████▇███████▇▇▇▆▇▇▆▆▅▇▆▅▄▅▅▄▄▆▅▅▅▆▄▄▄▅▄▅▆▆▆▆▆▅▆▆▅ █
  512 ns        Histogram: log(frequency) by time        886 ns <

 Memory estimate: 1.38 KiB, allocs estimate: 9.
```

83.7 x faster :tada: :tada: :tada: :tada: